feat: allow env variable for mqtt prefix
It is possible to configure a different prefix for homeassistant's auto
discover than the default of homeassistant. This allows a user to set
this as an environment variable if they have done so.

Uses `homeassistant/` by default, but if you pass `MQTT_TOPIC_PREFIX` as
a env variable we use that prefix instead